backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
motor
PyPDF2
pytesseract
openai
langchain
passlib[bcrypt]
python-jose
pandas
openpyxl
pdf2image
transformers
torch